dom/datastore/tests/test_bug976311.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test for DataStore - bug 976311</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<div id="container"></div>
  <script type="application/javascript;version=1.7">

  var gHostedManifestURL = 'http://test/tests/dom/datastore/tests/file_app.sjs?testToken=file_bug976311.html&template=file_bug976311.template.webapp';
  var gHostedManifestURL2 = 'http://example.com/tests/dom/datastore/tests/file_app.sjs?testToken=file_bug976311.html&template=file_bug976311.template.webapp';
  var gApps = [];

  function cbError() {
    ok(false, "Error callback invoked");
    finish();
  }

  function installApp(aApp) {
    var request = navigator.mozApps.install(aApp);
    request.onerror = cbError;
    request.onsuccess = function() {
      gApps.push(request.result);
      runTest();
    }
  }

  function uninstallApps() {
    if (!gApps.length) {
      ok(true, "All done!");
      runTest();
      return;
    }

    var app = gApps.pop();
    var request = navigator.mozApps.mgmt.uninstall(app);
    request.onerror = cbError;
    request.onsuccess = uninstallApps;
  }


  function testApp() {
   var ifr = document.createElement('iframe');
   ifr.setAttribute('mozbrowser', 'true');
   ifr.setAttribute('mozapp', gApps[0].manifestURL);
   ifr.setAttribute('src', gApps[0].manifest.launch_path);
   var domParent = document.getElementById('container');

   // Set us up to listen for messages from the app.
   var listener = function(e) {
     var message = e.detail.message;
     if (/^OK/.exec(message)) {
       ok(true, "Message from app: " + message);
     } else if (/KO/.exec(message)) {
       ok(false, "Message from app: " + message);
     } else if (/DONE/.exec(message)) {
       ok(true, "Messaging from app complete");
       ifr.removeEventListener('mozbrowsershowmodalprompt', listener);
       domParent.removeChild(ifr);
       runTest();
     }
   }

   // This event is triggered when the app calls "alert".
   ifr.addEventListener('mozbrowsershowmodalprompt', listener, false);
   domParent.appendChild(ifr);

  }

  var tests = [
    // Permissions
    function() {
      SpecialPowers.pushPermissions(
        [{ "type": "browser", "allow": 1, "context": document },
         { "type": "embed-apps", "allow": 1, "context": document },
         { "type": "webapps-manage", "allow": 1, "context": document }], runTest);
    },

    // Preferences
    function() {
      SpecialPowers.pushPrefEnv({"set": [["dom.datastore.enabled", true],
                                         ["dom.datastore.sysMsgOnChangeShortTimeoutSec", 1],
                                         ["dom.datastore.sysMsgOnChangeLongTimeoutSec", 3],
                                         ["dom.testing.ignore_ipc_principal", true],
                                         ["dom.testing.datastore_enabled_for_hosted_apps", true]]}, runTest);
    },

    function() {
      if (SpecialPowers.isMainProcess()) {
        SpecialPowers.Cu.import("resource://gre/modules/DataStoreChangeNotifier.jsm");
      }

      SpecialPowers.setAllAppsLaunchable(true);
      SpecialPowers.setBoolPref("dom.mozBrowserFramesEnabled", true);
      runTest();
    },

    // No confirmation needed when an app is installed
    function() {
      SpecialPowers.autoConfirmAppInstall(runTest);
    },

    // Installing the app1
    function() { installApp(gHostedManifestURL); },

    // Installing the app2
    function() { installApp(gHostedManifestURL2); },

    // Run tests in app
    testApp,

    // Uninstall the apps
    uninstallApps
  ];

  function runTest() {
    if (!tests.length) {
      finish();
      return;
    }

    var test = tests.shift();
    test();
  }

  function finish() {
    SimpleTest.finish();
  }

  SimpleTest.waitForExplicitFinish();
  runTest();
  </script>
</body>
</html>
